Update 17

See merge request ethereum-freeton-bridge/bridge-contracts!27

+++ free-ton/contracts/additional/EventEmitter.sol
@@ -3,7 +3,7 @@ pragma AbiHeader expire;
 
 
 contract EventEmitter {
-    uint static _randomNonce;
+    uint16 static _randomNonce;
 
     uint state;
